backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
pymysql
python-dotenv
celery
redis
requests
google-auth